// src/components/UI/Footer/enum.ts
export enum FooterClassNames {
  FOOTER = 'footer',
  CONTAINER = 'container',
  FOOTER_CONTENT = 'footer-content',
  GH_LINK = 'gh-link',
  RS_LINK = 'rs-link',
  FOOTER_IMG = 'footer-img',
  CREATED = 'created',
}

export enum FooterLinks {
  GH_LINK = 'https://github.com/TheMrCrowley',
  RS_LINK = 'https://rs.school/js/',
}